pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dnspython>=2.7.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.6",
    "pydantic>=2.10.5",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
]
